Fix an issue with GenerateJSON module, where it would insert empty keys

+++ .bin/index.js
@@ -4,6 +4,7 @@ var commander = require("commander");
 var shell = require("shelljs");
 var Promise = require("bluebird");
 var util_1 = require("./util");
+var Exception_1 = require("./Exception");
 var pkg = require('../package.json');
 var util_2 = require("./util");
 var CreateFile_1 = require("./CreateFile");
@@ -23,7 +24,9 @@ var command_generateJson = function (srcDir, options) {
         return CreateFile_1.createFile(options.out, json);
     })
         .then(util_1.passThrough(function () {
-        console.log('Successfully generated JSON files from', srcDir, 'at', options.out);
+        if (options.out) {
+            console.log('Successfully generated JSON files from', srcDir, 'at', options.out);
+        }
     }));
 };
 var command_generateTypes = function (jsonFilePath, options) {
@@ -39,7 +42,9 @@ var command_generateTypes = function (jsonFilePath, options) {
         return util_2.writeFile(options.out, tsd);
     }))
         .then(util_1.passThrough(function () {
-        console.log('Successfully generated Typescript .tsd from', jsonFilePath, 'at', options.out);
+        if (options.out) {
+            console.log('Successfully generated Typescript .tsd from', jsonFilePath, 'at', options.out);
+        }
     }));
 };
 commander
@@ -92,7 +97,9 @@ var command_generateClientSDK = function (appName, options) {
             .then(function () { return undefined; });
     })
         .then(util_1.passThrough(function () {
-        console.log('Successfully generated SDKs at', options.out);
+        if (options.out) {
+            console.log('Successfully generated SDKs at', options.out);
+        }
     }));
 };
 var untrackedFiles = function () {
@@ -104,14 +111,30 @@ var applyVersion = function (AppName, repoPath) {
     return Promise
         .resolve(getReleaseTypeFromFiles(tmp + "/" + AppName + ".json", compiled + "/" + AppName + ".json"))
         .then(function (releaseType) {
+        if (releaseType === 'none') {
+            return Promise.reject(new Exception_1.NoChangesException());
+        }
         return releaseType;
     })
         .then(util_1.passThrough(function (releaseType) {
+        if (untrackedFiles()) {
+            return Promise.reject(new Exception_1.UncommitedChanges());
+        }
         shell.rm('-rf', compiled);
         shell.mkdir(compiled);
         shell.cp('-R', tmp + "/*", compiled);
         shell.rm('-rf', tmp);
-    }));
+    }))
+        .then(util_1.passThrough(function () {
+        shell.exec("git add " + compiled);
+        shell.exec("git commit -m 'Beetlejuice Commit: Source Compiled.'");
+    }))
+        .then(function (releaseType) {
+        return shell.exec("npm version " + releaseType);
+    })
+        .then(function () {
+        return shell.exec('git push origin master; git push --tags');
+    });
 };
 commander
     .command('generate-client-sdks <AppName>')
